Mehdi_Rajabi_Mini_Project_1_Maktab60/requirements.txt
asgiref==3.4.1
Django==3.2.9
django-extensions==3.1.5
djangorestframework==3.12.4
flake8==4.0.1
importlib-metadata==4.2.0
mccabe==0.6.1
model-bakery==1.3.3
Pillow==8.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pygraphviz==1.6
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.1
zipp==3.6.0
asgiref==3.4.1
Django==3.2.9
django-extensions==3.1.5
django-uuslug==1.2.0
djangorestframework==3.12.4
flake8==4.0.1
importlib-metadata==4.2.0
mccabe==0.6.1
model-bakery==1.3.3
Pillow==8.4.0
pycodestyle==2.8.0
pyflakes==2.4.0
pygraphviz==1.6
python-slugify==5.0.2
pytz==2021.3
sqlparse==0.4.2
text-unidecode==1.3
typing_extensions==4.0.1
zipp==3.6.0
